{"schema":"decodex/commit/1","summary":"Modularize operator status projections","authority":"manual"}
