+++ apps/decodex/src/orchestrator/status/post_review/retry_budget.rs
@@ -1,0 +1,188 @@
+#[allow(clippy::wildcard_imports)] use super::*;
+
+pub(in crate::orchestrator) fn retry_budget_exhausted_post_review_lane_classification<I>(
+	snapshot: &PostReviewLaneSnapshot,
+	project: &ServiceConfig,
+	workflow: &WorkflowDocument,
+	review_state_inspector: &I,
+	mut classification: PostReviewLaneClassification,
+) -> PostReviewLaneClassification
+where
+	I: PullRequestReviewStateInspector,
+{
+	if classification.pr_url.is_none() {
+		classification.pr_url =
+			snapshot.review_handoff.as_ref().map(|marker| marker.pr_url().to_owned());
+	}
+	if classification.pr_state.is_none()
+		&& let Some(review_state) =
+			retry_budget_exhausted_merged_review_state(snapshot, review_state_inspector)
+	{
+		classification = initial_post_review_lane_classification(&review_state);
+
+		apply_pre_orchestration_post_review_classification(
+			snapshot,
+			workflow,
+			&review_state,
+			&mut classification,
+		);
+	}
+	if merged_closeout_pending_classification(&classification)
+		&& worktree_has_no_tracked_changes(snapshot.worktree.worktree_path())
+	{
+		if snapshot.issue.state.name == workflow.frontmatter().tracker().resolved_completed_state()
+			&& !worktree_has_no_tracked_changes(project.repo_root())
+		{
+			classification.decision = PostReviewLaneDecision::CleanupBlocked;
+			classification.reason = String::from("default_branch_worktree_dirty");
+
+			return classification;
+		}
+
+		return classification;
+	}
+	if classification.pr_state.as_deref() == Some("MERGED")
+		&& worktree_has_no_tracked_changes(snapshot.worktree.worktree_path())
+	{
+		classification.decision = if snapshot.issue.state.name
+			== workflow.frontmatter().tracker().resolved_completed_state()
+		{
+			PostReviewLaneDecision::CleanupBlocked
+		} else {
+			PostReviewLaneDecision::CloseoutBlocked
+		};
+		classification.reason = String::from("retry_budget_exhausted");
+
+		return classification;
+	}
+
+	classification.decision = PostReviewLaneDecision::Block;
+	classification.reason = String::from("retry_budget_exhausted");
+
+	classification
+}
+
+pub(in crate::orchestrator) fn merged_closeout_pending_classification(
+	classification: &PostReviewLaneClassification,
+) -> bool {
+	classification.decision == PostReviewLaneDecision::Continue
+		&& classification.reason == "pull_request_merged_closeout_pending"
+		&& classification.pr_state.as_deref() == Some("MERGED")
+}
+
+pub(in crate::orchestrator) fn confirm_status_visible_merged_closeout(
+	snapshot: &PostReviewLaneSnapshot,
+	project: &ServiceConfig,
+	classification: &mut PostReviewLaneClassification,
+) {
+	if !merged_closeout_pending_classification(classification) {
+		return;
+	}
+
+	let Some(pr_url) = classification.pr_url.as_deref() else {
+		mark_merged_closeout_confirmation_conflict(classification, None, None);
+
+		return;
+	};
+	let expected_head_sha = snapshot
+		.review_handoff
+		.as_ref()
+		.map(ReviewHandoffMarker::pr_head_oid)
+		.or(classification.pr_head_sha.as_deref());
+	let Some(expected_head_sha) = expected_head_sha else {
+		mark_merged_closeout_confirmation_conflict(classification, None, None);
+
+		return;
+	};
+	let github_token = match resolve_configured_env_var(
+		"github.token_env_var",
+		Some(project.github().token_env_var()),
+	) {
+		Ok(github_token) => github_token,
+		Err(error) => {
+			let root_cause = classify_pull_request_readback_report(&error);
+
+			mark_merged_closeout_confirmation_conflict(classification, None, Some(root_cause));
+
+			return;
+		},
+	};
+	let merge_readback = match github::inspect_pull_request_merge_readback(
+		snapshot.worktree.worktree_path(),
+		pr_url,
+		&github_token,
+		project.github().command_path(),
+	) {
+		Ok(merge_readback) => merge_readback,
+		Err(error) => {
+			let root_cause = classify_pull_request_readback_report(&error);
+
+			mark_merged_closeout_confirmation_conflict(classification, None, Some(root_cause));
+
+			return;
+		},
+	};
+
+	if merge_readback.state == "MERGED"
+		&& merge_readback.head_ref_oid.as_deref() == Some(expected_head_sha)
+	{
+		return;
+	}
+
+	mark_merged_closeout_confirmation_conflict(
+		classification,
+		Some(merge_readback),
+		Some(PullRequestReadbackRootCause::LineageValidationFailed),
+	);
+}
+
+pub(in crate::orchestrator) fn mark_merged_closeout_confirmation_conflict(
+	classification: &mut PostReviewLaneClassification,
+	merge_readback: Option<PullRequestMergeViewResponse>,
+	root_cause: Option<PullRequestReadbackRootCause>,
+) {
+	classification.decision = PostReviewLaneDecision::Block;
+	classification.reason = String::from("pull_request_merge_state_conflict");
+	classification.readback_warning = Some(String::from("pull_request_merge_state_conflict"));
+	classification.readback_root_cause =
+		root_cause.map(|root_cause| root_cause.as_str().to_owned());
+
+	if let Some(merge_readback) = merge_readback {
+		classification.pr_state = Some(merge_readback.state);
+		classification.pr_head_sha =
+			merge_readback.head_ref_oid.or_else(|| classification.pr_head_sha.clone());
+	}
+}
+
+pub(in crate::orchestrator) fn retry_budget_exhausted_merged_review_state<I>(
+	snapshot: &PostReviewLaneSnapshot,
+	review_state_inspector: &I,
+) -> Option<PullRequestReviewState>
+where
+	I: PullRequestReviewStateInspector,
+{
+	let review_handoff = snapshot.review_handoff.as_ref()?;
+
+	if !worktree_has_no_tracked_changes(snapshot.worktree.worktree_path()) {
+		return None;
+	}
+
+	let review_state = review_state_inspector
+		.inspect_review_state(snapshot.worktree.worktree_path(), review_handoff.pr_url())
+		.ok()?;
+
+	(review_state.state == "MERGED").then_some(review_state)
+}
+
+pub(in crate::orchestrator) fn worktree_has_no_tracked_changes(worktree_path: &Path) -> bool {
+	let Ok(output) = Command::new("git")
+		.arg("-C")
+		.arg(worktree_path)
+		.args(["status", "--porcelain", "--untracked-files=no"])
+		.output()
+	else {
+		return false;
+	};
+
+	output.status.success() && String::from_utf8_lossy(&output.stdout).trim().is_empty()
+}
